+++ scripts/schema/generate.sh
@@ -5,4 +5,4 @@ set -e
 cd "$(dirname "$0")";
 cd ../..
 
-cargo run --features="generate_schema" -- --generate_schema > schema/nightly/bottom.json
+cargo run --bin schema --features="generate_schema" -- > schema/nightly/bottom.json
