+++ src/bin/main.rs
@@ -1,0 +1,5 @@
+use bottom::start_bottom;
+
+fn main() -> anyhow::Result<()> {
+    start_bottom()
+}
