+++ src/options/args.rs
@@ -566,10 +566,6 @@ pub struct OtherArgs {
 
     #[arg(short = 'V', long, action = ArgAction::Version, help = "Prints version information.")]
     version: (),
-
-    #[cfg(feature = "generate_schema")]
-    #[arg(long, action = ArgAction::SetTrue)]
-    pub generate_schema: bool,
 }
 
 /// Returns a [`BottomArgs`].
